+++ src/molecules/tanstack-table/fields/tanstack-table-table-actions-custom-dialog.tsx
@@ -12,7 +12,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog';
-import { TanstackTableActionsCustomDialog } from '..';
+import { TanstackTableActionsCustomDialog } from '../types';
 
 type TanstackTableCustomDialogProps = {
   setDialogOpen: () => void;
